Code review comment fixes.
